requirements.txt
greenlet==3.1.1
playwright==1.49.1
pyee==12.0.0
PyYAML==6.0.2
typing_extensions==4.12.2
Flask==3.0.2
